Abstract filter component so it can be moved to a separate module
Start using redux middleware to update the URL path

change-type: minor
