+++ public/index.html
@@ -47,7 +47,6 @@
         schema: 'schema.yaml',
         // The attribute path to the documents content
         contentPath: 'Pokedex',
-        defaultView: 'Global view',
         repo: {
           name: 'pensieve',
           account: 'resin-io',
